@@ -1,15 +1,15 @@
 {
   "title": "ScrollController",
   "name": "properjs-scrollcontroller",
-  "description": "Window scroll event controller using request animation frame.",
-  "version": "0.2.2",
+  "description": "Overkill scroll handling for professionals.",
+  "version": "0.3.0",
   "homepage": "https://github.com/ProperJS/ScrollController",
   "readmeFilename": "README.md",
   "contributors": [
     {
-      "name": "Brandon Kitajchuk",
-      "email": "kitajchuk@gmail.com",
-      "url": "http://kitajchuk.com"
+      "name": "Brandon Lee Kitajchuk",
+      "email": "bk@kitajchuk.com",
+      "url": "https://kitajchuk.com"
     }
   ],
   "repository": {
@@ -27,11 +27,11 @@
   ],
   "main": "ScrollController",
   "scripts": {
-    "build": "./node_modules/.bin/webpack -d --colors",
-    "watch": "./node_modules/.bin/webpack -d --watch --colors"
+    "test": "./node_modules/.bin/webpack --progress --profile --display-modules --colors",
+    "start": "npm run test ; node_modules/.bin/webpack-dev-server"
   },
   "dependencies": {
-    "properjs-controller": "~0.2"
+    "properjs-controller": ">=0.3.0"
   },
   "keywords": [
     "properjs",
@@ -41,8 +41,15 @@
     "window"
   ],
   "devDependencies": {
-    "browser-sync": "^2.18.8",
-    "browser-sync-webpack-plugin": "^1.1.4",
-    "webpack": "^2.2.1"
+    "@babel/cli": "^7.8.4",
+    "@babel/core": "^7.9.0",
+    "@babel/polyfill": "^7.8.7",
+    "@babel/preset-env": "^7.9.0",
+    "babel-loader": "^8.1.0",
+    "eslint": "^7.2.0",
+    "eslint-loader": "^4.0.2",
+    "webpack": "^4.42.0",
+    "webpack-cli": "^3.3.11",
+    "webpack-dev-server": "^3.11.0"
   }
 }
